refactor: [051-refactor] 투표 로드 메서드 분리
