+++ src/main/java/project/votebackend/controller/vote/VoteLoadController.java
@@ -11,10 +11,10 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import project.votebackend.dto.vote.LoadVoteDto;
 import project.votebackend.security.CustumUserDetails;
+import project.votebackend.service.vote.VoteLoadService;
 import project.votebackend.service.vote.VoteService;
 import project.votebackend.util.PageResponseUtil;
 
-import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -22,7 +22,7 @@ import java.util.Map;
 @RequestMapping("/vote")
 public class VoteLoadController {
 
-    private final VoteService voteService;
+    private final VoteLoadService voteLoadService;
 
     //메인페이지 투표 불러오기 (자신이 작성한, 자신이 선택한 카테고리, 자신이 팔로우한 사람의 글)
     @GetMapping("/load-main-page-votes")
@@ -32,7 +32,7 @@ public class VoteLoadController {
             @RequestParam(defaultValue = "10") int size
     ) {
         Pageable pageable = PageRequest.of(page, size, Sort.by("createdAt").descending());
-        Page<LoadVoteDto> votePage = voteService.getMainPageVotes(userDetails.getId(), pageable);
+        Page<LoadVoteDto> votePage = voteLoadService.getMainPageVotes(userDetails.getId(), pageable);
         return ResponseEntity.ok(PageResponseUtil.toResponse(votePage));
     }
 
@@ -42,19 +42,8 @@ public class VoteLoadController {
             @PathVariable Long voteId,
             @AuthenticationPrincipal CustumUserDetails userDetails
     ) {
-        LoadVoteDto voteDto = voteService.getVoteById(voteId, userDetails.getId());
+        LoadVoteDto voteDto = voteLoadService.getVoteById(voteId, userDetails.getId());
         return ResponseEntity.ok(voteDto);
-    }
-
-    //좋아요 수 상위 게시물
-    //추후 수정해야할 api
-    @GetMapping("/top-liked")
-    public ResponseEntity<List<LoadVoteDto>> getTopLikedVotes(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @RequestParam(defaultValue = "30") int size) {
-
-        List<LoadVoteDto> result = voteService.getTopLikedVotes(userDetails.getUsername(), size);
-        return ResponseEntity.ok(result);
     }
 
     //특정 카테고리의 게시물 불러오기
@@ -65,7 +54,7 @@ public class VoteLoadController {
             @RequestParam(defaultValue = "30") int size,
             @AuthenticationPrincipal UserDetails userDetails
     ) {
-        Page<LoadVoteDto> votePage = voteService.getVotesByCategorySortedByLike(categoryId, page, size, userDetails.getUsername());
+        Page<LoadVoteDto> votePage = voteLoadService.getVotesByCategorySortedByLike(categoryId, page, size, userDetails.getUsername());
         return ResponseEntity.ok(PageResponseUtil.toResponse(votePage));
     }
 }
